# [Bugfix] Fix shape mismatch assertion error when loading Gemma3n model with BitsAndBytes quantization